00-install: update permissions management
Installing the system from scratch can take a while, I don't want to
have to monitor too closely what is happening but there is some
commands which must be run as my user and not root.

To do this I decided to provide my user access as root to all commands
without having to enter a password for 2 hours from the time the script
is started.