20-bspwm: remove nerd-fonts-hack in favor of ttf-font
I had a strange behavior during the last install.
I had only one of the two, the nerd one.
But I still didn't have the correcct icons in polybar.
I installed the ttf-hack one and it worked.
But with the cache and all I don't really understand how everything is
working so... I will try with a fresh install