20-bspwm.sh: install nerd fonts symbols
At the end it seems that the ttf-hack font didn't have the awesome
symbols.
Awesome symbols were properly rendered because there was no character in
the ttf-hack font to take their place, so the system fall-back to the
awesome font.

But with nerd-fonts-hack, which aggregates symbols from different
sources, there was different symbols for the same code.
Therefore they were used instead of the one from awesome.

So I decided to install ttf-nerd-fonts-symbols to get some symbols I
used from them (mainly in vim statusline) without having hack twice.

The last issue was that ttf-hack was not able to show some of the
symbols for my rofi powermenu.
It seems that the code of these symbols exist in ttf-hack but are empty,
quite annoying.
So I had to force the police to "Font Awesome 5 Free 10" in this rofi
script to get them shown correctly.